requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-socketio==5.11.2
requests==2.32.3
redis==5.0.1
aiohttp>=3.9.0
aiokafka>=0.12.0
httpx>=0.27.0
prometheus-client>=0.16.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
